requirements.txt
Flask~=3.1.1
Flask-SQLAlchemy
python-dotenv~=1.1.0
Authlib~=1.6.0
psycopg2-binary
SQLAlchemy~=2.0.41
Flask-Login~=0.6.3
Flask-Migrate~=4.1.0
alembic~=1.16.1